Se crear archivo functions.js para pasar las funciones que habia en index.js
